Consolidate boilerplate for test generation
